+++ include/wifi_battle_tower/app_state.h
@@ -22,6 +22,66 @@
 #include "string_gf.h"
 #include "string_template.h"
 
+enum WifiBattleTowerState {
+    BT_STATE_INITIAL_PROMPT = 0,
+    BT_STATE_INITIAL_YES_NO_MENU,
+    BT_STATE_INIT_INET,
+    BT_STATE_WAIT_INET_CONNECT,
+    BT_STATE_INIT_NAS_LOGIN,
+    BT_STATE_WAIT_NAS_LOGIN,
+    BT_STATE_INIT_HTTP,
+    BT_STATE_REQUEST_INFO,
+    BT_STATE_WAIT_INFO,
+    BT_STATE_SET_PROFILE,
+    BT_STATE_WAIT_SET_PROFILE,
+    BT_STATE_REQUEST_ROOM_COUNT,
+    BT_STATE_WAIT_ROOM_COUNT,
+    BT_STATE_SELECT_OPPONENT_PROMPT,
+    BT_STATE_DRAW_OPPONENT_SELECT,
+    BT_STATE_SELECT_OPPONENT,
+    BT_STATE_DOWNLOAD_PROMPT,
+    BT_STATE_REQUEST_DOWNLOAD,
+    BT_STATE_WAIT_DOWNLOAD,
+    BT_STATE_DOWNLOAD_COMPLETE_PROMPT,
+    BT_STATE_ALREADY_CHALLENGED_PROMPT,
+    BT_STATE_EXIT_PROMPT,
+    BT_STATE_EXIT_YES_NO,
+    BT_STATE_EXIT_MATCH_LIST_PROMPT,
+    BT_STATE_EXIT_MATCH_LIST_YES_NO,
+    BT_STATE_UPLOAD_PROMPT,
+    BT_STATE_REQUEST_UPLOAD,
+    BT_STATE_WAIT_UPLOAD,
+    BT_STATE_UPLOAD_COMPLETE_PROMPT,
+    BT_STATE_SELECT_RANK_PROMPT,
+    BT_STATE_DRAW_RANK_SELECT,
+    BT_STATE_SELECT_RANK,
+    BT_STATE_MATCH_LIST_SELECT_PROMPT,
+    BT_STATE_DRAW_MATCH_LIST_SELECT,
+    BT_STATE_SELECT_MATCH_LIST,
+    BT_STATE_DOWNLOAD_MATCH_LIST_PROMPT,
+    BT_STATE_REQUEST_MATCH_LIST,
+    BT_STATE_WAIT_MATCH_LIST,
+    BT_STATE_MATCH_LIST_COMPLETE_PROMPT,
+    BT_STATE_INIT_SAVE,
+    BT_STATE_WAIT_SAVE,
+    BT_STATE_GOODBYE_PROMPT,
+    BT_STATE_CLEANUP_INET,
+    BT_STATE_DISCONNECTING_PROMPT,
+    BT_STATE_IDLE,
+    BT_STATE_DISCONNECT_ENTRY,
+    BT_STATE_END_SESSION,
+    BT_STATE_WAIT_TEXT,
+    BT_STATE_WAIT_TEXT_YES_NO,
+    BT_STATE_WAIT_TEXT_TIMED,
+    BT_STATE_DISCONNECT_YES_NO,
+    BT_STATE_CONNECTED_PROMPT,
+    BT_STATE_CONNECTED_YES_NO_MENU,
+    BT_STATE_SHOW_ERROR,
+    BT_STATE_ERROR_CLEANUP,
+    BT_STATE_SHOW_DWC_ERROR,
+    BT_STATE_DWC_ERROR_ACK,
+};
+
 typedef struct WifiBattleTowerAppState {
     UnkStruct_0206BC70 *args;
     BgConfig *bgConfig;
@@ -30,8 +90,8 @@ typedef struct WifiBattleTowerAppState {
     int screenMode;
     int exitMode;
     int unused_18;
-    int state;
-    int nextState;
+    enum WifiBattleTowerState state;
+    enum WifiBattleTowerState nextState;
     void *dwcHeapBuffer;
     NNSFndHeapHandle dwcHeap;
     DWCInetControl inetControl;
