+++ src/battle_frontier_save.c
@@ -4,13 +4,12 @@
 
 #include "struct_defs/battle_frontier.h"
 
-#include "wifi_battle_tower/save.h"
-
 #include "battle_hall_save.h"
 #include "savedata.h"
 #include "unk_0202FF4C.h"
 #include "unk_020302D0.h"
 #include "unk_02030494.h"
+#include "wifi_battle_tower_save.h"
 #include "wifi_list.h"
 
 int BattleFrontierSave_Size(void)
