+++ src/overlay104/frscrcmd.c
@@ -48,7 +48,6 @@
 #include "overlay104/struct_ov104_0223F174_decl.h"
 #include "overlay104/struct_ov104_0223F1B4.h"
 #include "savedata/save_table.h"
-#include "wifi_battle_tower/save.h"
 
 #include "battle_frontier.h"
 #include "battle_frontier_stats.h"
@@ -89,6 +88,7 @@
 #include "unk_02033200.h"
 #include "unk_020363E8.h"
 #include "vars_flags.h"
+#include "wifi_battle_tower_save.h"
 #include "wifi_list.h"
 
 #include "constdata/const_020EA358.h"
