+++ src/unk_0209ACF4.c
@@ -5,7 +5,6 @@
 
 #include "field/field_system.h"
 #include "overlay005/fieldmap.h"
-#include "wifi_battle_tower/save.h"
 
 #include "bg_window.h"
 #include "colored_arrow.h"
@@ -24,6 +23,7 @@
 #include "system.h"
 #include "text.h"
 #include "unk_0203D1B8.h"
+#include "wifi_battle_tower_save.h"
 
 #include "res/text/bank/easy_chat.h"
